move spritebuf to seprate file
